refactor(user): move user query methods from UserService to UserQueryService; update usages across controllers and services
