+++ chatbot/src/main/java/pt/estga/chatbot/services/ChatbotAccountFacade.java
@@ -4,9 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import pt.estga.chatbot.events.ChatbotAccountConnectedEventPublisher;
-import pt.estga.user.entities.User;
 import pt.estga.user.services.ChatbotAccountService;
-import pt.estga.user.services.UserService;
+import pt.estga.user.services.UserQueryService;
 
 /**
  * Facade encapsulating the logic to link a messaging platform identity to a domain user
@@ -18,13 +17,13 @@ import pt.estga.user.services.UserService;
 public class ChatbotAccountFacade {
 
     private final ChatbotAccountService chatbotAccountService;
-    private final UserService userService;
+    private final UserQueryService userQueryService;
     private final ChatbotAccountConnectedEventPublisher eventPublisher;
 
     public boolean linkPlatformIdentity(Long domainUserId, String platform, String platformUserId) {
         if (domainUserId == null) return false;
 
-        return userService.findById(domainUserId).map(user -> {
+        return userQueryService.findById(domainUserId).map(user -> {
             try {
                 chatbotAccountService.createOrUpdateChatbot(user, platformUserId);
                 eventPublisher.publish(platform, platformUserId, user.getId());
