+++ user/src/main/java/pt/estga/user/controllers/AdminUserController.java
@@ -52,7 +52,7 @@ public class AdminUserController {
     public ResponseEntity<UserDto> getById(
             @Parameter(description = "ID of the user to be retrieved", required = true)
             @PathVariable Long id) {
-        return service.findById(id)
+        return queryService.findById(id)
                 .map(mapper::toDto)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
