+++ user/src/main/java/pt/estga/user/services/UserQueryService.java
@@ -3,6 +3,7 @@ package pt.estga.user.services;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import pt.estga.sharedweb.filtering.QueryProcessor;
 import pt.estga.sharedweb.models.PagedRequest;
 import pt.estga.sharedweb.models.QueryResult;
@@ -10,6 +11,8 @@ import pt.estga.user.dtos.UserDto;
 import pt.estga.user.entities.User;
 import pt.estga.user.mappers.UserMapper;
 import pt.estga.user.repositories.UserRepository;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -28,5 +31,27 @@ public class UserQueryService {
         );
 
         return userPage.map(mapper::toDto);
+    }
+
+    public Optional<User> findById(Long id) {
+        return repository.findById(id);
+    }
+
+    @Transactional(readOnly = true)
+    public Optional<User> findByEmail(String email) {
+        return repository.findByEmail(email);
+    }
+
+    @Transactional(readOnly = true)
+    public Optional<User> findByKeycloakSub(String keycloakSub) {
+        return repository.findByKeycloakSub(keycloakSub);
+    }
+
+    public Optional<User> findByIdForProfile(Long id) {
+        return repository.findByIdForProfile(id);
+    }
+
+    public boolean existsByUsername(String username) {
+        return repository.existsByUsername(username);
     }
 }
