+++ user/src/main/java/pt/estga/user/services/UserService.java
@@ -16,28 +16,6 @@ public class UserService {
     private final UserRepository repository;
     private final ChatbotAccountRepository chatbotAccountRepository;
 
-    public Optional<User> findById(Long id) {
-        return repository.findById(id);
-    }
-
-    @Transactional(readOnly = true)
-    public Optional<User> findByEmail(String email) {
-        return repository.findByEmail(email);
-    }
-
-    @Transactional(readOnly = true)
-    public Optional<User> findByKeycloakSub(String keycloakSub) {
-        return repository.findByKeycloakSub(keycloakSub);
-    }
-
-    public Optional<User> findByIdForProfile(Long id) {
-        return repository.findByIdForProfile(id);
-    }
-
-    public boolean existsByUsername(String username) {
-        return repository.existsByUsername(username);
-    }
-
     public User create(User user) {
         return repository.save(user);
     }
